[Chat] Mention refactor part 3 (#3149) Migrate mention functionality to TextFieldWithMention component
