+++ change-beta/@azure-communication-react-7af99641-b8f3-4327-a80c-2c41e059e8db.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "Refactor mention out into a new component",
+  "packageName": "@azure/communication-react",
+  "email": "3941071+emlynmac@users.noreply.github.com",
+  "dependentChangeType": "none"
+}
